resume-screening-ai/requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
PyPDF2
docx2txt
nltk
requests
python-dotenv
